# src/rest/docs/swagger.json
{
	"basePath": "/api/v2",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"ConfigModel": {
			"description": "UI config",
			"properties": {
				"domain": {
					"type": "string",
					"description": "domain this app is running"
				},
				"facebook_app_id": {
					"type": "string",
					"description": "facebook SlackMap id"
				},
				"facebook_scope": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"required": [
				"domain"
			],
			"type": "object"
		},
		"LocationPathModel": {
			"properties": {},
			"type": "object"
		},
		"MeModel": {
			"description": "User model definition",
			"properties": {
				"rid": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"facebook_id": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"imperial": {
					"type": "boolean"
				},
				"location_path": {
					"$ref": "#/definitions/LocationPathModel"
				},
				"firstname": {
					"type": "string"
				},
				"lastname": {
					"type": "string"
				},
				"login_at": {
					"type": "string"
				}
			},
			"type": "object"
		},
		"MeGetResponse": {
			"properties": {
				"me": {
					"$ref": "#/definitions/MeModel"
				}
			},
			"required": [
				"me"
			],
			"type": "object"
		},
		"AuthLoginByFbResponse": {
			"properties": {
				"me": {
					"$ref": "#/definitions/MeModel"
				}
			},
			"required": [
				"me"
			],
			"type": "object"
		},
		"AuthLoginByFbRequest": {
			"properties": {
				"accessToken": {
					"type": "string"
				},
				"signedRequest": {
					"type": "string"
				}
			},
			"required": [
				"accessToken",
				"signedRequest"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "@slackmap/api",
		"version": "0.0.1",
		"description": "SlackMap API Server",
		"license": {
			"name": "GPL-3.0"
		}
	},
	"paths": {
		"/config": {
			"get": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ConfigModel"
						}
					}
				},
				"description": "current application configuration object",
				"tags": [
					"slackmap"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/me": {
			"get": {
				"operationId": "MeGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/MeGetResponse"
						}
					}
				},
				"description": "current application configuration object",
				"tags": [
					"slackmap"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/loginByFb": {
			"post": {
				"operationId": "LoginByFb",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/AuthLoginByFbResponse"
						}
					}
				},
				"description": "current application configuration object",
				"tags": [
					"slackmap"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "data",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AuthLoginByFbRequest"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_token": {
			"type": "apiKey",
			"name": "api_token",
			"in": "query",
			"description": ""
		}
	},
	"host": "localhost:3000"
}